File restructure and theme change
Changed the file structure of i3, i3status and Xresources. They are now divided in config-main and config-theme, for easier theming. In each config folder is a bash file for auto "glue" the 2 files. The new theme is a very rought attempt to mimic base16 style of breeze dark, will revisit in the future.